refactor: token, 유저 api 분리
