+++ src/app/(auth)/mypage/page.tsx
@@ -6,22 +6,10 @@ import { IoChevronBackSharp } from 'react-icons/io5';
 import axiosInstance from '@/app/api/axiosInstance';
 import { LuLogOut } from 'react-icons/lu';
 import { useQuery } from '@tanstack/react-query';
-import { fetchLoginData } from '@/utils/fetchLoginData';
+import { getUserData } from '@/services/user/getUserData';
 
 const MyPage = () => {
   const router = useRouter();
-
-  const { isLoading, error, data } = useQuery({
-    queryKey: ['loginData'],
-    queryFn: fetchLoginData,
-  });
-  if (!isLoading) {
-    console.log(data);
-  }
-
-  if (error) {
-    console.log(error);
-  }
 
   const handleGoBack = () => {
     router.back();
@@ -29,18 +17,27 @@ const MyPage = () => {
 
   const handleLogout = async () => {
     try {
-      const response = await axiosInstance.post('api/logout');
+      const response = await axiosInstance.post('/api/logout');
 
       if (response.status === 200) {
         localStorage.setItem('isLoggedIn', 'false');
-        alert('로그아웃 되었습니다!');
+        alert('로그아웃되었습니다.');
         router.push('/');
       }
     } catch (logoutError) {
       console.log('로그아웃 실패: ', logoutError);
-      alert('죄송합니다. 로그아웃에 실패했습니다. 잠시 후 다시 시도해주세요.');
+      alert('죄송합니다. 로그아웃에 실패했습니다. 잠시 후 다시 시도해 주세요.');
     }
   };
+
+  const { error, data: user } = useQuery({
+    queryKey: ['loginData'],
+    queryFn: getUserData,
+  });
+
+  if (error) {
+    console.log(error);
+  }
 
   return (
     <div className="md:h-100 sm:h-100 relative h-full w-full p-5">
@@ -59,9 +56,9 @@ const MyPage = () => {
         <IoChevronBackSharp size={24} />
       </div>
       <div>
-        <p>{data?.nickname}</p>
-        <p>{data?.socialType}</p>
-        <p>{data?.email}</p>
+        <p>{user?.nickname}</p>
+        <p>{user?.socialType}</p>
+        <p>{user?.email}</p>
       </div>
       <button type="submit" className="flex gap-2" onClick={handleLogout}>
         <LuLogOut aria-label="로그아웃" size={24} />
