+++ src/services/user/getUserData.ts
@@ -1,21 +1,17 @@
 import axiosInstance from '@/app/api/axiosInstance';
-import { UserInfo } from '@/types/userInfo';
-import LocalStorage from './localstorage';
+import { UserInfo } from '@/types/user/userInfo';
+import LocalStorage from '../../utils/localstorage';
 
-export const fetchLoginData = async (): Promise<UserInfo | null> => {
+export const getUserData = async (): Promise<UserInfo | null> => {
   const isLoggedIn = LocalStorage.getItem('isLoggedIn');
   if (isLoggedIn === 'true') {
-    const result = await axiosInstance.patch('/api/tokens/refresh');
+    const result = await axiosInstance.get('/api/user/my-info');
     const user = result.data.results;
-    const accessToken = result.headers.authorization;
-
-    axiosInstance.defaults.headers.common.Authorization = `Bearer ${accessToken}`;
 
     const userData: UserInfo = {
       email: user.email,
       socialType: user.socialType,
       id: user.id,
-      // firstLoginCheck: user.firstLoginCheck,
       nickname: user.nickname,
       imageUrl: user.imageUrl,
       selfIntroduction: user.selfIntroduction,
@@ -26,10 +22,7 @@ export const fetchLoginData = async (): Promise<UserInfo | null> => {
       mateCompleteParticipationCount: user.mateCompleteParticipationCount,
       teamMatchingCompleteParticipationCount:
         user.teamMatchingCompleteParticipationCount,
-      accessToken,
     };
-
-    console.log(userData);
     return userData;
   }
   return null;
